backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0
pandas==2.2.2
google-cloud-bigquery>=3.11.0
google-auth>=2.23.0
requests>=2.32.0
pdfplumber>=0.11.4
pyarrow>=17.0.0
db-dtypes>=1.2.0